requirements.txt
discord==2.3.2
nox==2025.5.1
ruff==0.11.13
black==25.1.0
feedparser==6.0.11
coverage==7.9.0
python-dotenv==1.1.0